fix(p24): psalm Batch 2 continued - array-access safety, ~20 more files
Fixes InvalidArrayOffset/PossiblyUndefinedArrayOffset/EmptyArrayAccess
across CatModifyPageRenderer.php, InstallationStats.php, PwgBase32.php,
NotificationByMailSubController.php, CommentsController.php,
ContainerDetector.php, HistoryService.php, and PictureCommentRenderer.php,
using the extract-once/guard idioms established throughout this batch.

Real fixes: PwgBase32.php's base_convert() return needed an (int) cast
before indexing the int-keyed $map array (base_convert always returns
string). ContainerDetector.php's `$file_lines[count($file_lines) - 1]`
could index at -1 on an empty file() result -- added a real non-empty
guard. Two more COUNT(*)/aggregate-query list-destructures replaced
with indexed `?? null` reads (CatModifyPageRenderer.php,
InstallationStats.php), same pattern as the several fixed earlier in
this batch.

The remaining ~15 sites across AlbumNotificationPageRenderer.php,
CatListPageRenderer.php, ElementSetRanksPageRenderer.php, Lang.php,
FilterService.php, CategoryCatsRenderer.php, SiteUpdateSubController.php,
PermalinksSubController.php, AlbumsPageRenderer.php,
UserListPageRenderer.php, and FeatureFlag.php all hit the same classes
of PHPStan-vs-Psalm disagreement documented throughout this batch
(guard-then-bare-access shape narrowing, by-ref narrowing loss on
dynamically-built trees, dynamic-include $conf/require patterns) --
each confirmed clean via composer analyse:phpstan with zero source
changes, covered by documented psalm.xml suppressions instead of
further casts.

Full unit/arch suite + PHPStan + ECS all green.

+++ src/Piwigo/Auth/PwgBase32.php
@@ -82,7 +82,7 @@ class PwgBase32
         $base32 = '';
         $i = 0;
         while ($i < count($fiveBitBinaryArray)) {
-            $base32 .= self::$map[base_convert(str_pad($fiveBitBinaryArray[$i], 5, '0'), 2, 10)];
+            $base32 .= self::$map[(int) base_convert(str_pad($fiveBitBinaryArray[$i], 5, '0'), 2, 10)];
             $i++;
         }
         if ($padding && ($x = strlen($binaryString) % 40) !== 0) {
